# openapi.json
{
  "components": {
    "schemas": {
      "AcceptanceLedgerSummary": {
        "description": "WOG1-F3 marked-provisional bridge: the ON-DISK\n`acceptance_ledger.json` (written by\n`acceptance.acceptance_ledger_bytes`) is a DIFFERENT shape than the\nin-memory `regolith.orchestrator.acceptance.AcceptanceOutcome`\nmodel it is rendered from (no `accepted_hashes` field on disk, and\n`accepted_deviations` replaces `deviations`) -- there is no\nregolith model that round-trips this file directly, so\n`AcceptanceOutcome.model_validate_json` fails against it. This is a\ngraphite-local shape mirroring exactly what ships to disk, escalated\nto the lithos coordinator as a gap (a `LedgerDocument` model in\n`regolith.orchestrator.acceptance` would close it).",
        "properties": {
          "accepted_deviations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AcceptedDeviation"
            },
            "title": "Accepted Deviations",
            "type": "array"
          },
          "cli_accepts_used": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cli Accepts Used",
            "type": "array"
          },
          "errors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "refusals": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Refusals",
            "type": "array"
          }
        },
        "title": "AcceptanceLedgerSummary",
        "type": "object"
      },
...
      "AcceptedDeviation": {
        "description": "One row of `dist/acceptance_ledger.json`'s `accepted_deviations`\narray (WOG1-F3 below) -- field-for-field what\n`regolith.orchestrator.acceptance.acceptance_ledger_bytes` writes.",
        "properties": {
          "accepted": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Accepted",
            "type": "array"
          },
          "basis": {
            "title": "Basis",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Digest"
          },
          "expires": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "match_set": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Match Set",
            "type": "array"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "basis",
          "kind"
        ],
        "title": "AcceptedDeviation",
        "type": "object"
      },
...
      "AuditIndex": {
        "description": "The package-level audit index: total accounting + per-obligation rows.",
        "properties": {
          "project": {
            "title": "Project",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AuditRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/AuditSummary"
          }
        },
        "required": [
          "project",
          "summary",
          "rows"
        ],
        "title": "AuditIndex",
        "type": "object"
      },
      "AuditRow": {
        "description": "One obligation's disposition in the audit index (D221.2).\n\n``detail`` names the exact evidence: for a ``calc_sheet`` the sheet\nid; for an ``accepted_deviation`` the waiver target + its memo digest\n(cross-linking ``acceptance_ledger.json``, never duplicating it); for\na ``deferred``/``violated`` row the named reason.",
        "properties": {
          "claim_name": {
            "title": "Claim Name",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "disposition": {
            "enum": [
              "calc_sheet",
              "accepted_deviation",
              "deferred",
              "violated"
            ],
            "title": "Disposition",
            "type": "string"
          },
          "subject_anchor": {
            "title": "Subject Anchor",
            "type": "string"
          }
        },
        "required": [
          "claim_name",
          "subject_anchor",
          "content_hash",
          "disposition",
          "detail"
        ],
        "title": "AuditRow",
        "type": "object"
      },
      "AuditSummary": {
        "description": "The audit index's obligation accounting (D221.2).\n\nTWO honest denominators, both reported so they can never be confused:\n\n* the CENSUS-shape counts (``discharged`` = results with no deferral;\n  ``accepted_deviation`` = the number of UNIQUE accepted obligation\n  content addresses -- exactly ``len(acceptance.accepted_hashes)``,\n  the same value ``fleet._census_from_report`` records; ``violated``)\n  so the audit index reconciles with ``fleet_census.json`` field for\n  field via :meth:`census_row`;\n* the ROW partition (``discharged`` + ``accepted_rows`` + ``deferred``\n  + ``violated`` == ``obligations``), because a calc sheet and an\n  audit row are PER OBLIGATION and forall-expanded obligation\n  instances legitimately share one content address -- so the number\n  of accepted ROWS (:attr:`accepted_rows`) can exceed the number of\n  unique accepted addresses (:attr:`accepted_deviation`). The row\n  partition is the zero-unexplained property (:meth:`balanced`).",
        "properties": {
          "accepted_deviation": {
            "title": "Accepted Deviation",
            "type": "integer"
          },
          "accepted_rows": {
            "title": "Accepted Rows",
            "type": "integer"
          },
          "deferred": {
            "title": "Deferred",
            "type": "integer"
          },
          "discharged": {
            "title": "Discharged",
            "type": "integer"
          },
          "obligations": {
            "title": "Obligations",
            "type": "integer"
          },
          "violated": {
            "title": "Violated",
            "type": "integer"
          }
        },
        "required": [
          "obligations",
          "discharged",
          "accepted_deviation",
          "accepted_rows",
          "deferred",
          "violated"
        ],
        "title": "AuditSummary",
        "type": "object"
      },
...
      "CalcInput": {
        "description": "One input to a calc sheet's obligation, with its provenance pin.\n\n``value`` is the datum as written (unit-carrying text is kept\nverbatim -- the source is the truth, never re-normalized); ``pin`` is\nthe content address for a ``record_ref`` (else empty). ``source`` is\nthe human-readable origin (the record name, the literal expression,\nor the derived value-source text).",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pin": {
            "default": "",
            "title": "Pin",
            "type": "string"
          },
          "provenance": {
            "enum": [
              "record_ref",
              "declared_literal",
              "derived",
              "unresolved"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "name",
          "value",
          "provenance"
        ],
        "title": "CalcInput",
        "type": "object"
      },
      "CalcSheet": {
        "description": "One discharged obligation's calc sheet (D221.1).",
        "properties": {
          "attestation": {
            "title": "Attestation",
            "type": "string"
          },
          "chain": {
            "$ref": "#/components/schemas/EvidenceChain"
          },
          "citation": {
            "title": "Citation",
            "type": "string"
          },
          "claim_name": {
            "title": "Claim Name",
            "type": "string"
          },
          "claim_text": {
            "title": "Claim Text",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/CalcInput"
            },
            "title": "Inputs",
            "type": "array"
          },
          "margin": {
            "title": "Margin",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "sheet_id": {
            "title": "Sheet Id",
            "type": "string"
          },
          "solver": {
            "title": "Solver",
            "type": "string"
          },
          "subject_anchor": {
            "title": "Subject Anchor",
            "type": "string"
          },
          "subject_ref": {
            "title": "Subject Ref",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "sheet_id",
          "claim_name",
          "claim_text",
          "subject_anchor",
          "subject_ref",
          "model_id",
          "model_version",
          "citation",
          "solver",
          "tier",
          "attestation",
          "inputs",
          "value",
          "margin",
          "verdict",
          "chain"
        ],
        "title": "CalcSheet",
        "type": "object"
      },
...
      "EvidenceChain": {
        "description": "The content-hash chain proving a calc sheet's derivation (D221.1).\n\n``sheet_digest`` is the producer-local blake3 of the sheet's own\ncanonical bytes (``local-blake3:`` tagged, charter 38 sec. 1.4);\n``evidence_hash`` / ``subject_ref`` / ``record_pins`` are the\nCANONICAL Rust content addresses off the payload (untagged). A\nreviewer follows sheet -> evidence -> payload -> sources without\ntrusting anything but the shipped bytes.",
        "properties": {
          "evidence_hash": {
            "title": "Evidence Hash",
            "type": "string"
          },
          "payload_refs": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Payload Refs",
            "type": "array"
          },
          "record_pins": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Record Pins",
            "type": "array"
          },
          "sheet_digest": {
            "title": "Sheet Digest",
            "type": "string"
          },
          "subject_ref": {
            "title": "Subject Ref",
            "type": "string"
          }
        },
        "required": [
          "sheet_digest",
          "evidence_hash",
          "subject_ref"
        ],
        "title": "EvidenceChain",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "graphite backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project}/acceptance-ledger": {
      "get": {
        "description": "The accepted-deviation ledger (waiver/memo panel, project-view\ndeliverable 2): every accepted deviation with its memo evidence\ndigest and basis text, read verbatim (WOG1-F3 provisional bridge).",
        "operationId": "get_acceptance_ledger_api_projects__project__acceptance_ledger_get",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptanceLedgerSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Acceptance Ledger",
        "tags": [
          "calc"
        ]
      }
    },
...
    "/api/projects/{project}/calc/audit": {
      "get": {
        "description": "The project's audit index (summary + rows) -- the same model\n`/obligations` filters/groups, exposed here unfiltered for a raw\nview (04.1 \"ANY DETAIL VIEW\" floor: raw-JSON toggle).",
        "operationId": "get_audit_index_api_projects__project__calc_audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Index",
        "tags": [
          "calc"
        ]
      }
    },
    "/api/projects/{project}/calc/sheets": {
      "get": {
        "description": "Every calc sheet in the project's shipped calc book (04.1 \"ANY\nTABLE\" floor: sort/filter/CSV-export are frontend concerns over\nthis same array, not a server-side feature).",
        "operationId": "list_calc_sheets_api_projects__project__calc_sheets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalcSheet"
                  },
                  "title": "Response List Calc Sheets Api Projects  Project  Calc Sheets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Calc Sheets",
        "tags": [
          "calc"
        ]
      }
    },
...
  }
}